final refactor with dimensions in cube and multicube
