+++ aggregator/src/test/scala/com/stratio/sparkta/aggregator/CubeSpec.scala
@@ -18,7 +18,7 @@ package com.stratio.sparkta.aggregator
 
 import java.io.{Serializable => JSerializable}
 
-import com.stratio.sparkta.plugin.dimension.native._
+import com.stratio.sparkta.plugin.dimension.default._
 import org.apache.spark.streaming.TestSuiteBase
 import org.joda.time.DateTime
 import org.junit.runner.RunWith
